refactor: event-log-api module
