+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/projects/delete/ELProjectFinder.scala
@@ -21,9 +21,9 @@ package io.renku.knowledgegraph.projects.delete
 import cats.MonadThrow
 import cats.effect.Async
 import cats.syntax.all._
+import io.renku.eventlog.api.EventLogClient
+import io.renku.eventlog.api.EventLogClient.{Result, SearchCriteria}
 import io.renku.events.consumers.Project
-import io.renku.graph.eventlog.EventLogClient
-import io.renku.graph.eventlog.EventLogClient.{Result, SearchCriteria}
 import io.renku.graph.model.events.EventInfo
 import io.renku.graph.model.projects
 import io.renku.http.rest.paging.model.PerPage
