+++ webhook-service/src/test/scala/io/renku/webhookservice/hookcreation/HookCreatorSpec.scala
@@ -21,13 +21,13 @@ package io.renku.webhookservice.hookcreation
 import cats.effect.IO
 import cats.effect.std.Queue
 import cats.syntax.all._
+import io.renku.eventlog
+import io.renku.eventlog.api.events.CommitSyncRequest
 import io.renku.events.consumers.ConsumersModelGenerators.consumerProjects
 import io.renku.events.consumers.Project
 import io.renku.generators.CommonGraphGenerators._
 import io.renku.generators.Generators.Implicits._
 import io.renku.generators.Generators._
-import io.renku.graph.eventlog
-import io.renku.graph.eventlog.api.events.CommitSyncRequest
 import io.renku.graph.model.projects.GitLabId
 import io.renku.http.client.AccessToken
 import io.renku.interpreters.TestLogger
